Move parse_class_name to utils
Move parse_class_name out of config where it does not belong, and
add it to utils. Adjust imports accordingly. Also move the unit
test to a new file test_utils.py.
