+++ test/test_config.py
@@ -1,4 +1,5 @@
 import unittest
+
 import sbol3
 
 
@@ -20,21 +21,6 @@ class TestConfig(unittest.TestCase):
         base_uri = 'https://synbiohub.org'
         sbol3.set_namespace(base_uri)
         self.assertEqual(base_uri, sbol3.get_namespace())
-
-    def test_parse_class_name(self):
-        # Test parsing with # delimiter in type URI
-        c = sbol3.Component('c', sbol3.SBO_DNA)
-        class_name = sbol3.config.parse_class_name(c.type_uri)
-        self.assertEqual(class_name, 'Component')
-
-        # Test parsing with # delimiter in type URI
-        m = sbol3.Measure(0, 'dollars')
-        class_name = sbol3.config.parse_class_name(m.type_uri)
-        self.assertEqual(class_name, 'Measure')
-
-        # Test failure with invalid type URI
-        with self.assertRaises(ValueError):
-            sbol3.config.parse_class_name('Component')
 
 
 if __name__ == '__main__':
